+++ src/main/java/taskBoard/exeption/BoardNotFoundException.java
@@ -2,7 +2,7 @@ package taskBoard.exeption;
 
 public class BoardNotFoundException extends AppException {
 
-    public BoardNotFoundException(Long id) {
+    public BoardNotFoundException(Integer id) {
         super("Board with id: " + id + " not found.");
     }
