+++ src/main/java/taskBoard/repository/VersionReleaseRepository.java
@@ -3,5 +3,5 @@ package taskBoard.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import taskBoard.model.VersionRelease;
 
-public interface VersionReleaseRepository extends JpaRepository<VersionRelease, Long> {
+public interface VersionReleaseRepository extends JpaRepository<VersionRelease, Integer> {
 }
